Fix HRM batch: configurable customer/invoice fields, POS decimals, row click, auto-download
- Customer list rows are now clickable across all cells, not just the name.
- POS field accepts decimals with both '.' and ',' separators; pos column
  widened from Int to Float.
- Customer system fields (name/email/phone/status/address) and invoice
  system fields (invoice number/date/payment terms/...) are upserted on
  configuration load and surfaced in their tabs as locked system rows
  (relabel + required togglable; key/type/delete locked).
- New "Invoice Details" configuration tab with two sub-tabs:
  - Additional Invoice Fields (system + custom invoice fields)
  - Variables (line item fields, placeholders, custom report variables)
- After a document is generated it now also auto-downloads in the browser
  in addition to being saved to the per-customer and global Documents tabs.

+++ src/lib/validators/index.ts
@@ -56,7 +56,7 @@ export const reportTemplateSchema = z.object({
 
 export const lineItemSchema = z.object({
   id: z.string().optional(),
-  pos: z.number().int(),
+  pos: z.number(),
   key: z.string().optional().nullable(),
   label: z.string().optional().nullable(),
   description: z.string().optional().nullable(),
